+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Edit playground functionality
+- Admin panel
+
+### Changed
+- Improve city determination
 
 ## [0.12.8] - 2025-03-08